pyproject.toml
[project]
name = "polymarket-analysis"
version = "0.1.0"
description = "Polymarket crypto trading analysis for statistical arbitrage"
authors = [
    {name = "User"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "jupyter (>=1.1.1,<2.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "numpy (>=2.3.0,<3.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "scipy (>=1.15.3,<2.0.0)",
    "scikit-learn (>=1.7.0,<2.0.0)",
    "plotly (>=6.1.2,<7.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "yfinance (>=0.2.63,<0.3.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "aiohttp (>=3.12.13,<4.0.0)",
    "asyncio-throttle (>=1.0.2,<2.0.0)",
    "py-clob-client (>=0.23.0,<1.0.0)",
    "spans (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
jupyter = "^1.1.1"
notebook = "^7.4.3"

